+++ source/TimeSeries/TimeSeriesBundleIngestor/Monitor/Databricks/DefaultTags.cs
@@ -1,0 +1,10 @@
+﻿namespace Energinet.DataHub.TimeSeries.TimeSeriesBundleIngestor.Monitor.Databricks;
+
+public class DefaultTags
+{
+    public string? ClusterName { get; set; }
+
+    public string? JobId { get; set; }
+
+    public string? RunName { get; set; }
+}
