+++ source/TimeSeries/TimeSeriesBundleIngestor/Monitor/HealthCheckEndpoint.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.Core.App.FunctionApp.Diagnostics.HealthChecks;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 
-namespace Energinet.DataHub.TimeSeries.TimeSeriesBundleIngestor
+namespace Energinet.DataHub.TimeSeries.TimeSeriesBundleIngestor.Monitor
 {
     public class HealthCheckEndpoint
     {
